% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScSpTITH.r
\name{ScSpTITHscore}
\alias{ScSpTITHscore}
\title{ScSpTITH: A Rank-Correlation Framework for Robust Quantification of Multi-Dimensional Tumor Heterogeneity (ScSpTITHscore)}
\usage{
ScSpTITHscore(
  data,
  meta,
  mode = c("overall", "cluster"),
  patient_col = "Patient",
  condition_col = "condition",
  top_n_genes = 5000
)
}
\arguments{
\item{data}{Gene expression matrix with rows as genes and columns as cells/samples}

\item{meta}{Metadata containing at minimum:
#'     - First column: cell/sample IDs (must match column names in data)
- Patient ID column (required)
- Condition/cluster column (optional, only required for cluster mode)}

\item{mode}{Calculation mode: "overall" - calculate for entire sample;
"cluster" - calculate separately for each cluster}

\item{patient_col}{Column name in metadata containing patient IDs, default "Patient"}

\item{condition_col}{Column name in metadata containing cluster/group information, default "condition"}

\item{top_n_genes}{Number of highly variable genes used for correlation calculation, default 5000}
}
\value{
A data frame containing Patient,ScSpTITHscore, Cluster (mode:"cluster")
}
\description{
ScSpTITH: A Rank-Correlation Framework for Robust Quantification of Multi-Dimensional Tumor Heterogeneity (ScSpTITHscore)
}
\examples{
# Mode 1: Calculate ITHscore for entire sample
#result1 <- ScSpTITHscore(data, meta, mode = "overall",patient_col = "Patient")
# Mode 2: Calculate ITHscore separately for each cluster
#result2 <- ScSpTITHscore(data_count, meta, mode = "cluster",patient_col = "Patient",condition_col = "condition")
}
